+++ common/src/main/java/io/github/mortuusars/envelope/client/gui/screen/AddressTagScreen.java
@@ -3,6 +3,7 @@ package io.github.mortuusars.envelope.client.gui.screen;
 import com.mojang.blaze3d.platform.InputConstants;
 import io.github.mortuusars.envelope.Envelope;
 import io.github.mortuusars.envelope.client.gui.widget.AddressBoxSuggestions;
+import io.github.mortuusars.envelope.world.mail.service.ServiceAddresses;
 import io.github.mortuusars.mortaar.client.Minecrft;
 import io.github.mortuusars.envelope.util.Colors;
 import io.github.mortuusars.envelope.world.mail.address.Address;
@@ -124,7 +125,7 @@ public class AddressTagScreen extends Screen {
               knownAddresses.blocks(),
               knownAddresses.players(),
               knownAddresses.services().stream()
-                    .filter(address -> !address.getDefinitionHolder().is(Envelope.Tags.ServiceAddresses.HIDDEN))
+                    .filter(address -> ServiceAddresses.isEnabled(address) && !address.isHidden())
                     .collect(Collectors.toSet()));
     }