+++ common/src/main/java/io/github/mortuusars/envelope/client/state/FillRecipientState.java
@@ -1,6 +1,0 @@
-package io.github.mortuusars.envelope.client.state;
-
-public class FillRecipientState {
-    public String recipient = "";
-    public String sender = "";
-}