+++ common/src/main/java/io/github/mortuusars/envelope/world/mail/service/cloud_depository/CloudDepository.java
@@ -1,9 +1,9 @@
 package io.github.mortuusars.envelope.world.mail.service.cloud_depository;
 
 import com.mojang.logging.LogUtils;
+import io.github.mortuusars.envelope.Config;
 import io.github.mortuusars.envelope.Envelope;
 import io.github.mortuusars.envelope.util.Colors;
-import io.github.mortuusars.envelope.world.item.SealedItem;
 import io.github.mortuusars.envelope.world.item.component.mail.log.DeliveryRecord;
 import io.github.mortuusars.envelope.world.item.component.seal.Seal;
 import io.github.mortuusars.envelope.world.item.component.seal.SealMaterial;
@@ -17,11 +17,9 @@ import io.github.mortuusars.envelope.world.mail.service.ServiceAddresses;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.RegistryAccess;
 import net.minecraft.core.component.DataComponents;
-import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.HoverEvent;
-import net.minecraft.network.chat.MutableComponent;
-import net.minecraft.network.chat.Style;
+import net.minecraft.network.chat.*;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.Mth;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.NotNull;
@@ -32,19 +30,24 @@ import java.util.List;
 import java.util.ListIterator;
 
 public class CloudDepository {
-    private static final Logger LOGGER = LogUtils.getLogger();
+    protected static final Logger LOGGER = LogUtils.getLogger();
 
     public static final ResourceLocation LETTER_MEANING_WITHDRAWAL_REQUEST = Envelope.resource("cloud_depository/withdrawal_request");
     public static final ResourceLocation LETTER_MEANING_STATUS_REQUEST = Envelope.resource("cloud_depository/status_request");
+    public static final ResourceLocation LETTER_MEANING_EXPANSION_REQUEST = Envelope.resource("cloud_depository/expansion_request");
 
     public static final Component RETURN_MESSAGE_NO_IDENTITY = Component.translatable(
           "gui.envelope.delivery_log.message.cloud_depository.no_identity").withColor(Colors.TOOLTIP_RED);
+    public static final Component RETURN_MESSAGE_NON_SERVICEABLE_IDENTITY = Component.translatable(
+          "gui.envelope.delivery_log.message.cloud_depository.non_serviceable_identity").withColor(Colors.TOOLTIP_RED);
     public static final Component RETURN_MESSAGE_STORAGE_IS_FULL = Component.translatable(
           "gui.envelope.delivery_log.message.cloud_depository.storage_is_full").withColor(Colors.TOOLTIP_RED);
     public static final Component RETURN_MESSAGE_STORAGE_IS_EMPTY = Component.translatable(
           "gui.envelope.delivery_log.message.cloud_depository.storage_is_empty").withColor(Colors.TOOLTIP_RED);
     public static final Component RETURN_MESSAGE_INDEX_NOT_FOUND = Component.translatable(
           "gui.envelope.delivery_log.message.cloud_depository.index_not_found").withColor(Colors.TOOLTIP_RED);
+    public static final Component RETURN_MESSAGE_CAPACITY_ALREADY_AT_MAX = Component.translatable(
+          "gui.envelope.delivery_log.message.cloud_depository.capacity_already_at_max").withColor(Colors.TOOLTIP_RED);
     public static final Component LOG_MESSAGE_DEPOSITED = Component.translatable(
           "gui.envelope.delivery_log.message.cloud_depository.deposited").withColor(Colors.TOOLTIP_GREEN);
     public static final Component LOG_MESSAGE_WITHDRAWN = Component.translatable(
@@ -80,6 +83,14 @@ public class CloudDepository {
         return seal;
     }
 
+    public int getTotalItemCount() {
+        int count = 0;
+        for (CloudDepositoryData.Account accountData : getData().accounts().values()) {
+            count += accountData.getTotal();
+        }
+        return count;
+    }
+
     // --
 
     public static MailDropOffResult handleDropOff(MailDropOffContext context) {
@@ -87,8 +98,6 @@ public class CloudDepository {
     }
 
     public MailDropOffResult handleDelivery(MailDropOffContext context) {
-        //TODO: Config
-
         // Should be handled higher up the chain, but just to be safe:
         if (context.isReturned()) {
             LOGGER.info("'{}' received returned mail. Delivery: '{}'. Voiding.", context.getTarget(), context.getDelivery());
@@ -97,12 +106,21 @@ public class CloudDepository {
 
         ItemStack mail = context.getMail();
 
+        if (!Config.Server.SERVICE_CLOUD_DEPOSITORY_ENABLED.get()) {
+            return MailDropOffResult.returned(mail, DeliveryRecord.Message.UNAVAILABLE);
+        }
+
         if (!(mail.get(Envelope.DataComponents.SEAL) instanceof Seal mailSeal)
               || mailSeal.signature().getString().isEmpty()) {
             return MailDropOffResult.returned(mail, RETURN_MESSAGE_NO_IDENTITY);
         }
 
         String account = mailSeal.signature().getString();
+
+        if (Config.Server.SERVICE_CLOUD_DEPOSITORY_ONLY_PLAYERS.get()
+              && service.getKnownPlayers().getDataOf(account).isEmpty()) {
+            return MailDropOffResult.returned(mail, RETURN_MESSAGE_NON_SERVICEABLE_IDENTITY);
+        }
 
         if (mail.is(Envelope.Items.SEALED_LETTER.get())
               && mail.get(Envelope.DataComponents.LETTER_MEANING) instanceof ResourceLocation meaning) {
@@ -112,13 +130,16 @@ public class CloudDepository {
             if (meaning.equals(LETTER_MEANING_STATUS_REQUEST)) {
                 return handleStatusRequest(context, account);
             }
+            if (meaning.equals(LETTER_MEANING_EXPANSION_REQUEST)) {
+                return handleExpansionRequest(context, account);
+            }
             return MailDropOffResult.returned(mail, DeliveryRecord.Message.REJECTED);
         }
 
         return handleItemDeposit(context, account);
     }
 
-    private MailDropOffResult handleItemDeposit(MailDropOffContext context, String account) {
+    protected MailDropOffResult handleItemDeposit(MailDropOffContext context, String account) {
         ItemStack mail = context.getMail();
         CloudDepositoryData.Account accountData = getData().ofAccount(account);
 
@@ -140,19 +161,6 @@ public class CloudDepository {
         return MailDropOffResult.reply(reply);
     }
 
-    protected MailDropOffResult handleStatusRequest(MailDropOffContext context, String account) {
-        LOGGER.debug("Received status request from: {}", account);
-
-        CloudDepositoryData.Account accountData = getData().ofAccount(account);
-
-        ItemStack statusReportLetter = createStatusReportLetter(seal, account,
-              accountData.getTotal(), accountData.getCapacity(), accountData.getItems(), context.getLevel().getRandom());
-        return MailDropOffResult.reply(Mail.of(statusReportLetter)
-              .sender(getAddress())
-              .recipient(context.getOrigin())
-              .get());
-    }
-
     protected MailDropOffResult handleWithdrawalRequest(MailDropOffContext context, String account) {
         LOGGER.debug("Received withdrawal request from: {}", account);
 
@@ -171,6 +179,8 @@ public class CloudDepository {
                     iterator.remove();
                     getData().setDirty();
 
+                    LOGGER.debug("Withdrew indexed [{}] from: {}", account, item);
+
                     return MailDropOffResult.reply(Mail.of(item)
                           .sender(getAddress())
                           .recipient(context.getOrigin())
@@ -181,21 +191,57 @@ public class CloudDepository {
             return MailDropOffResult.returned(context.getMail(), RETURN_MESSAGE_INDEX_NOT_FOUND);
         }
 
-        ItemStack removedItem = accountData.getItems().removeLast();
+        ItemStack item = accountData.getItems().removeLast();
         getData().setDirty();
 
-        return MailDropOffResult.reply(Mail.of(removedItem)
+        LOGGER.debug("Withdrew [{}] from: {}", account, item);
+
+        return MailDropOffResult.reply(Mail.of(item)
               .sender(getAddress())
               .recipient(context.getOrigin())
               .writeToLog(DeliveryRecord.message(LOG_MESSAGE_WITHDRAWN)).get());
     }
 
-    public int getTotalItemCount() {
-        int count = 0;
-        for (CloudDepositoryData.Account accountData : getData().accounts().values()) {
-            count += accountData.getTotal();
+    protected MailDropOffResult handleStatusRequest(MailDropOffContext context, String account) {
+        LOGGER.debug("Received status request from: {}", account);
+
+        CloudDepositoryData.Account accountData = getData().ofAccount(account);
+
+        ItemStack report = createStatusReportLetter(seal, account, accountData.getTotal(), accountData.getCapacity(),
+              Config.Server.SERVICE_CLOUD_DEPOSITORY_ACCOUNT_STORAGE_MAX_CAPACITY.get(),
+              accountData.getItems(), context.getLevel().getRandom());
+        return MailDropOffResult.reply(Mail.of(report)
+              .sender(getAddress())
+              .recipient(context.getOrigin())
+              .get());
+    }
+
+    protected MailDropOffResult handleExpansionRequest(MailDropOffContext context, String account) {
+        LOGGER.debug("Received expansion request from: {}", account);
+
+        int maxCapacity = Config.Server.SERVICE_CLOUD_DEPOSITORY_ACCOUNT_STORAGE_MAX_CAPACITY.get();
+        int increase = Config.Server.SERVICE_CLOUD_DEPOSITORY_ACCOUNT_STORAGE_CAPACITY_PER_EXPANSION.get();
+
+        CloudDepositoryData.Account accountData = getData().ofAccount(account);
+        int oldCapacity = accountData.getCapacity();
+
+        if (accountData.getCapacity() >= maxCapacity) {
+            return MailDropOffResult.returned(context.getMail(), RETURN_MESSAGE_CAPACITY_ALREADY_AT_MAX);
         }
-        return count;
+
+        int newCapacity = Mth.clamp(accountData.getCapacity() + increase, 1, maxCapacity);
+        accountData.setCapacity(newCapacity);
+        getData().setDirty();
+
+        LOGGER.debug("Increased storage capacity of account '{}' to {}", account, maxCapacity);
+
+        int increasedBy = newCapacity - oldCapacity;
+
+        ItemStack report = createExpansionReportLetter(seal, account, accountData.getCapacity(), maxCapacity, increasedBy);
+        return MailDropOffResult.reply(Mail.of(report)
+              .sender(getAddress())
+              .recipient(context.getOrigin())
+              .get());
     }
 
     // --
@@ -224,14 +270,14 @@ public class CloudDepository {
         return Mail.createSealedLetter(Component.translatable("letter.envelope.cloud_depository.deposit_report.message",
                     depositedItemComponent,
                     withdrawalRequestComponent,
-                    Component.literal(accountStorageCurrent + "/" + accountStorageCapacity)))
+                    Component.literal(accountStorageCurrent + "/" + accountStorageCapacity).withStyle(ChatFormatting.DARK_RED)))
               .itemName(Component.translatable("letter.envelope.cloud_depository.deposit_report.name"))
               .set(Envelope.DataComponents.SEAL, seal)
               .get();
     }
 
     public static ItemStack createStatusReportLetter(Seal seal, String account, int accountStorageCurrent, int accountStorageCapacity,
-                                                     List<ItemStack> storedItems, RandomSource random) {
+                                                     int maxCapacity, List<ItemStack> storedItems, RandomSource random) {
         List<ItemStack> storedIndexableItems = storedItems.stream()
               .filter(i -> i.has(DataComponents.CUSTOM_NAME))
               .toList();
@@ -260,11 +306,32 @@ public class CloudDepository {
                   randomIndexableItems);
         }
 
+        Component expansion = accountStorageCapacity < maxCapacity
+              ? Component.translatable("letter.envelope.cloud_depository.status_report.expansion",
+                    Component.literal(Integer.toString(maxCapacity)).withStyle(ChatFormatting.DARK_RED))
+              : CommonComponents.EMPTY;
+
         return Mail.createSealedLetter(Component.translatable("letter.envelope.cloud_depository.status_report.message",
                     Component.literal(account).withStyle(ChatFormatting.DARK_RED),
                     Component.literal(accountStorageCurrent + "/" + accountStorageCapacity).withStyle(ChatFormatting.DARK_RED),
-                    indexable))
+                    indexable, expansion))
               .itemName(Component.translatable("letter.envelope.cloud_depository.status_report.name"))
+              .set(Envelope.DataComponents.SEAL, seal)
+              .get();
+    }
+
+    public static ItemStack createExpansionReportLetter(Seal seal, String account, int accountStorageCapacity, int maxCapacity, int increasedBy) {
+        Component further = accountStorageCapacity < maxCapacity
+              ? Component.translatable("letter.envelope.cloud_depository.expansion_report.further_expansion",
+                    Component.literal(Integer.toString(maxCapacity)).withStyle(ChatFormatting.DARK_RED))
+              : CommonComponents.EMPTY;
+
+        return Mail.createSealedLetter(Component.translatable("letter.envelope.cloud_depository.expansion_report.message",
+                    Component.literal(account).withStyle(ChatFormatting.DARK_RED),
+                    Component.literal(Integer.toString(increasedBy)).withStyle(ChatFormatting.DARK_RED),
+                    Component.literal(Integer.toString(accountStorageCapacity)).withStyle(ChatFormatting.DARK_RED),
+                    further))
+              .itemName(Component.translatable("letter.envelope.cloud_depository.expansion_report.name"))
               .set(Envelope.DataComponents.SEAL, seal)
               .get();
     }
@@ -278,6 +345,12 @@ public class CloudDepository {
     public static ItemStack createStatusRequestLetter() {
         return Mail.createLetter(Component.translatable("letter.envelope.cloud_depository.status_request.message"))
               .set(Envelope.DataComponents.LETTER_MEANING, CloudDepository.LETTER_MEANING_STATUS_REQUEST)
+              .get();
+    }
+
+    public static ItemStack createExpansionRequestLetter() {
+        return Mail.createLetter(Component.translatable("letter.envelope.cloud_depository.expansion_request.message"))
+              .set(Envelope.DataComponents.LETTER_MEANING, CloudDepository.LETTER_MEANING_EXPANSION_REQUEST)
               .get();
     }
 }